Utilities

A number of functions are made available to the user to retrieve specific components of the TFT solution

Quick Links

Utilities

Quick Links

Amplitude

Phase

Anti-Rotating Phase

Dynamic Phasor

Anti-Rotating Dynamic Phasor

Signal

Error

Amplitude

TFT.a — Function

TFT.a(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Shorthand function to obtain the amplitude of the Dth-degree derivative of the Hth-harmonic phasor, dispatching to amplitude(sol,D,H).

TFT.amplitude — Function

TFT.amplitude(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Function to obtain the amplitude of the Dth-degree derivative of the Hth-harmonic phasor.

aₕ⁽⁰⁾(t) = 2 ⋅ |ξₕ⁽⁰⁾(t)|
# aₕ⁽¹⁾(t) = 2 ⋅ ℜ[ξₕ⁽¹⁾(t) ⋅ exp(-im ⋅ ϕₕ⁽⁰⁾(t))] 
# aₕ⁽²⁾(t) = 2 ⋅ ℜ[ξₕ⁽²⁾(t) ⋅ exp(-im ⋅ ϕₕ⁽⁰⁾(t))] + aₕ⁽⁰⁾(t) ⋅ φₕ⁽¹⁾(t)^2

See: Assessing Synchrophasor Estimates of an Event Captured by a Phasor Measurement Unit, pg. 3112

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

D::Int | degree of the derivative [-], default=0

H::Int | harmonic number [-], default=1

Output:

a::Vector{<:Real} | amplitude aₕ⁽ᴰ⁾(t) [?]

Phase

TFT.ϕ — Function

TFT.ϕ(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Shorthand function to obtain the phase of the Dth-degree derivative of the Hth-harmonic phasor, dispatching to phase(sol, D, H).

TFT.phase — Function

TFT.phase(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Function to obtain the alternative angle of the Dth-degree derivative of the Hth-harmonic phasor.

ϕₕ⁽⁰⁾(t) = ∠[pₕ⁽⁰⁾(t)]

See: Assessing Synchrophasor Estimates of an Event Captured by a Phasor Measurement Unit, pg. 3112

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

H::Int | harmonic number [-], default=1

Output:

ϕ::Vector{<:Real} | phase ϕₕ⁽ᴰ⁾(t) [(rad)]

Anti-Rotating Phase

TFT.φ — Function

TFT.φ(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Shorthand function to obtain the anti-rotating phase of the Dth-degree derivative of the Hth-harmonic phasor, dispatching to angle(sol,D,H).

TFT.ar_phase — Function

TFT.ar_phase(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Function to obtain the anti-rotating phase of the Dth-degree derivative of the Hth-harmonic phasor.

φₕ⁽⁰⁾(t) = ∠[ξₕ⁽⁰⁾(t) ⋅ exp(-2.0 ⋅ im ⋅ pi ⋅ H / N) ⋅ 0:length(s(t))-1]
# φₕ⁽¹⁾(t) = ℑ[pₕ⁽¹⁾(t) ⋅ exp(-im ⋅ ϕₕ⁽⁰⁾(t))] / aₕ⁽⁰⁾(t)
# φₕ⁽²⁾(t) = (ℑ[pₕ⁽²⁾(t) ⋅ exp(-im ⋅ ϕₕ⁽⁰⁾(t))] + 2 ⋅ aₕ⁽¹⁾(t) ⋅ φₕ⁽¹⁾(t)) / aₕ⁽⁰⁾(t)

See: Assessing Synchrophasor Estimates of an Event Captured by a Phasor Measurement Unit, pg. 3112

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

D::Int | degree of the derivative [-], default=0

H::Int | harmonic number [-], default=1

Output:

φ::Vector{<:Real} | anti-rotating phase φₕ⁽ᴰ⁾(t) [(rad)]

Dynamic Phasor

TFT.ξ — Function

TFT.ξ(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Shorthand function to obtain the Dth-degree derivative of the Hth-harmonic dynamic phasor, dispatching to phasor(sol,D,H).

TFT.phasor — Function

TFT.phasor(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Function to obtain the D-th-degree derivative of the Hth-harmonic dynamic phasor.

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

D::Int | degree of the derivative [-], default=0

H::Int | harmonic number [-], default=1

Output:

phasor::Vector{<:Complex} | dynamic phasor ξₕ⁽ᴰ⁾(t) [?]

Anti-Rotating Dynamic Phasor

TFT.ψ — Function

TFT.ψ(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Shorthand function to obtain the Dth-degree derivative of the Hth-harmonic anti- rotating dynamic phasor, dispatching to ar_phasor(sol,D,H).

TFT.ar_phasor — Function

TFT.ar_phasor(sol::TFT.AbstractDTFTSolution, D::Int=0, H::Int=1)

Function to obtain the D-th-degree derivative of the Hth-harmonic anti-rotating dynamic phasor.

ψₕ⁽ᴰ⁾(t) = ξₕ⁽ᴰ⁾(t) exp(-im ωₕ t)

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

D::Int | degree of the derivative [-], default=0

H::Int | harmonic number [-], default=1

Output:

ar_phasor::Vector{<:Complex} | anti-rotating dynamic phasor ψₕ⁽ᴰ⁾(t) [?]

Signal

TFT.signal — Method

TFT.signal(sol::TFT.AbstractDTFTSolution)

Function to obtain the overall signal.

s(t) = ∑ₕ ξₕ⁽⁰⁾(t) + conj(ξₕ⁽⁰⁾(t))

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

Output:

signal::Vector{<:Real} | signal s(t) [?]

TFT.signal — Method

TFT.signal(sol::TFT.AbstractDTFTSolution, H::Int)

Function to obtain the Hth-harmonic signal.

sₕ(t) = ξₕ⁽⁰⁾(t) + conj(ξₕ⁽⁰⁾(t))

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

H::Int | harmonic number [-]

Output:

signal::Vector{<:Real} | signal sₕ(t) [?]

Error

TFT.error — Method

TFT.error(sol::TFT.AbstractDTFTSolution)

Function to obtain the error between the input and computed signal.

Input:

sol::AbstractDTFTSolution | DTFT solution struct [-]

Output:

e::Vector{<:Real} | error [?]